requirements.txt
aiogram
black
langchain-gigachat
python-dateutil
requests